Beat Maker

+++ Engine/Source/Vector2.h
@@ -17,8 +17,8 @@ struct Vector2
 
 	Vector2(int x, int y)
 	{
-		this -> x = x;
-		this -> y = y;
+		this -> x = (float)x;
+		this -> y = (float)y;
 	}
 
 	Vector2 operator + (Vector2 v) const
